Avoid bad setup on Windows???

+++ src/main.rs
@@ -753,10 +753,11 @@ fn build_map(data: &Vec<FitDataRecord>, ui: &UserInterface, mc_rc: &Rc<MapCache>
         .is_some()
     {
         let mc = &**mc_rc;
-        let source = libshumate::MapSourceRegistry::with_defaults()
-            .by_id("osm-mapnik")
-            .unwrap();
-        ui.map.set_map_source(Some(&source));
+        let source = libshumate::MapSourceRegistry::with_defaults().by_id("osm-mapnik");
+        if source.is_none() {
+            panic!("No map source found.")
+        }
+        ui.map.set_map_source(Some(&source.unwrap()));
         // Get values from fit file.
         let run_path = &mc.run_path;
         ui.path_layer.as_ref().unwrap().remove_all();
